# argus/docs/swagger.yaml
basePath: /
definitions:
...
  argus_app_internal_model.DetectionLineDirection:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-comments:
      DetectionLineDirectionAToB: 沿折线方向跨越
      DetectionLineDirectionBToA: 逆折线方向跨越
      DetectionLineDirectionBoth: 双向跨越
    x-enum-descriptions:
    - 双向跨越
    - 沿折线方向跨越
    - 逆折线方向跨越
    x-enum-varnames:
    - DetectionLineDirectionBoth
    - DetectionLineDirectionAToB
    - DetectionLineDirectionBToA
  argus_app_internal_model.DetectionPoint:
    properties:
      x:
        type: number
      "y":
        type: number
    type: object
  argus_app_internal_model.DetectionRule:
    properties:
      lineDirection:
        $ref: '#/definitions/argus_app_internal_model.DetectionLineDirection'
      points:
        items:
          $ref: '#/definitions/argus_app_internal_model.DetectionPoint'
        type: array
      role:
        $ref: '#/definitions/argus_app_internal_model.DetectionRuleRole'
    type: object
  argus_app_internal_model.DetectionRuleRole:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      DetectionRuleRoleLine: 分界线：目标跨线触发
      DetectionRuleRoleMask: 屏蔽/遮罩区域：此区域内目标忽略
      DetectionRuleRoleROI: 布防/检测区域：只在此区域内检测
      DetectionRuleRoleUnspecified: 未指定（非法）
    x-enum-descriptions:
    - 未指定（非法）
    - 布防/检测区域：只在此区域内检测
    - 屏蔽/遮罩区域：此区域内目标忽略
    - 分界线：目标跨线触发
    x-enum-varnames:
    - DetectionRuleRoleUnspecified
    - DetectionRuleRoleROI
    - DetectionRuleRoleMask
    - DetectionRuleRoleLine
...
  argus_app_internal_pkg_response.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  argus_app_internal_service.AlarmRecordDetail:
    properties:
      alarmTypeId:
        type: string
      algorithmId:
        type: string
      algorithmName:
        type: string
      algorithmVersion:
        type: string
      bbox:
        description: '[x1, y1, x2, y2]'
        items:
          type: number
        type: array
      cameraId:
        type: string
      cameraName:
        type: string
      confidence:
        type: number
      createdAt:
        type: string
      eventId:
        type: string
      id:
        type: integer
      imageId:
        type: string
      imageUrl:
        type: string
      instanceId:
        type: string
      occurredAt:
        type: string
      rules:
        items:
          $ref: '#/definitions/argus_app_internal_model.DetectionRule'
        type: array
      targetLabel:
        type: string
      timeSynced:
        type: boolean
      trackId:
        type: integer
    type: object
  argus_app_internal_service.AlarmRecordItem:
    properties:
      alarmTypeId:
        type: string
      algorithmId:
        type: string
      algorithmName:
        type: string
      algorithmVersion:
        type: string
      bbox:
        description: '[x1, y1, x2, y2]'
        items:
          type: number
        type: array
      cameraId:
        type: string
      cameraName:
        type: string
      confidence:
        type: number
      createdAt:
        type: string
      eventId:
        type: string
      id:
        type: integer
      imageId:
        type: string
      imageUrl:
        type: string
      instanceId:
        type: string
      occurredAt:
        type: string
      targetLabel:
        type: string
      timeSynced:
        type: boolean
      trackId:
        type: integer
    type: object
  argus_app_internal_service.AlarmRecordPageResult:
    properties:
      items:
        items:
          $ref: '#/definitions/argus_app_internal_service.AlarmRecordItem'
        type: array
      total:
        type: integer
    type: object
...
  argus_app_internal_service.LoginResult:
    properties:
      accessToken:
        type: string
      avatar:
        type: string
      desc:
        type: string
      homePath:
        type: string
      realName:
        type: string
      roles:
        items:
          type: string
        type: array
      userId:
        type: string
      username:
        type: string
    type: object
...
  argus_app_internal_service.UserInfoDTO:
    properties:
      avatar:
        type: string
      desc:
        type: string
      homePath:
        type: string
      realName:
        type: string
      roles:
        items:
          type: string
        type: array
      userId:
        type: string
      username:
        type: string
    type: object
...
  internal_api.AccessCodesResponse:
    properties:
      code:
        type: integer
      data:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
...
  internal_api.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_api.LoginResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_service.LoginResult'
      message:
        type: string
    type: object
...
  internal_api.UserInfoResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_service.UserInfoDTO'
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: argus 后端 API 接口文档
  title: argus API
  version: "1.0"
paths:
  /api/auth/codes:
    get:
      description: 获取当前登录用户拥有的全部按钮权限码
      produces:
      - application/json
      responses:
        "200":
          description: 权限码列表
          schema:
            $ref: '#/definitions/internal_api.AccessCodesResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取当前用户权限码集合
      tags:
      - 认证模块
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录获取 AccessToken，同时在 Cookie 中写入 RefreshToken
      parameters:
      - description: 登录参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回用户信息及 AccessToken
          schema:
            $ref: '#/definitions/internal_api.LoginResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      summary: 用户登录
      tags:
      - 认证模块
...
  /api/auth/refresh:
    post:
      description: 从 HttpOnly Cookie 中读取 RefreshToken，签发新的 AccessToken 并轮换 RefreshToken
        Cookie
      produces:
      - text/plain
      responses:
        "200":
          description: 新的 AccessToken
          schema:
            type: string
        "401":
          description: 未授权或 RefreshToken 无效
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      summary: 刷新令牌
      tags:
      - 认证模块
...
  /api/record/alarms:
    get:
      description: 支持按发生时间区间、摄像头、算法、告警类型、置信度区间组合分页查询
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页条数
        in: query
        name: pageSize
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      - description: 摄像头ID
        in: query
        name: cameraId
        type: string
      - description: 算法ID
        in: query
        name: algorithmId
        type: string
      - description: 告警类型ID
        in: query
        name: alarmTypeId
        type: string
      - description: 最低置信度
        in: query
        name: minConfidence
        type: number
      - description: 最高置信度
        in: query
        name: maxConfidence
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: 告警分页数据
          schema:
            allOf:
            - $ref: '#/definitions/argus_app_internal_pkg_response.Result'
            - properties:
                data:
                  $ref: '#/definitions/argus_app_internal_service.AlarmRecordPageResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 分页查询告警记录
      tags:
      - 告警记录
  /api/record/alarms/{id}:
    get:
      description: 获取单条告警详细信息，含检测规则与目标框
      parameters:
      - description: 告警记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 告警详情数据
          schema:
            allOf:
            - $ref: '#/definitions/argus_app_internal_pkg_response.Result'
            - properties:
                data:
                  $ref: '#/definitions/argus_app_internal_service.AlarmRecordDetail'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "404":
          description: 记录不存在
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 查询告警记录详情
      tags:
      - 告警记录
  /api/record/images/{id}:
    get:
      description: 校验 Bearer Token 并在后端代理读取 Engine var/images 下的安全图片流，支持 type=thumb
        缩略图
      parameters:
      - description: 图片唯一标识 image_id
        in: path
        name: id
        required: true
        type: string
      - description: '图片类型: thumb 缩略图, 默认原图'
        in: query
        name: type
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: 图片文件流
          schema:
            type: file
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "404":
          description: 图片不存在
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 受控读取告警图片
      tags:
      - 告警记录
...
  /api/user/info:
    get:
      description: 根据请求头中的 Bearer Token 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 当前用户信息
          schema:
            $ref: '#/definitions/internal_api.UserInfoResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取当前登录用户信息
      tags:
      - 认证模块
...
securityDefinitions:
  BearerAuth:
    description: 在 Header 中传入 `Bearer <token>` 进行身份认证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"